% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{matrixPlotter}
\alias{matrixPlotter}
\alias{GenomeMatrix}
\title{Plot upper half of a matrix rotated -45 degrees.}
\usage{
matrixPlotter(mat, granges = NULL, plotGR = NULL, extend = 0.5,
  heightProp = 1/3, zlim = NULL, colorBias = 1, highlight = NULL,
  colRamp = "YlOrRd", colPalette = NULL, smoothFilt = NULL)
}
\arguments{
\item{mat}{A n x n symmetric matrix to plot.}

\item{granges}{An optional genomic ranges object of length n representing the genomic coordinates the 
rows and columns of the matrix.}

\item{plotGR}{A \code{\link{GenomicRanges}} object indicating the genomic region to be plotted. 
This option is ignored if granges is not provided.}

\item{extend}{Proportion of the plotting region to extend both upstream and downstream. These 
extra regions are normally be plotted, but if white spaces appear in the upper corners of the 
plot, consider increasing the value of this parameter.}

\item{heightProp}{Proportion of the plotting region (x-axis) to plot in the y-axis.}

\item{zlim}{Limit for the maximum value of z (color) to plot. Larger values will be truncated
to this parameter.}

\item{colorBias}{Bias parameter for color palette (see ?colorRampPalette).}

\item{highlight}{Optional. A \code{\link{GenomicRanges}} object indicating a region to highlight. 
If specified, A box would be drawn around this region.}

\item{colRamp}{A name of a color palette (see ?brewer.pal).}

\item{colPalette}{A vector of colors in hexadecimal format. Overwrites the palette for \code{colRamp}.}

\item{smoothFilt}{An optional smoothing function for the matrix.}
}
\value{
A ggplot2 object.
}
\description{
This function inputs a matrix along with an optional \code{\link{GenomicRanges}}
object and plots the upper half of the matrix rotated -45 degrees.
}
\examples{
library(HiTC)
data(Dixon2012_IMR90, package="HiCDataHumanIMR90")
mat <- as.matrix( hic_imr90_40@.Data[[1]]@intdata )
granges <- hic_imr90_40@.Data[[1]]@xgi
pl <- matrixPlotter( log2(mat+1), granges, plotGR=GRanges("chr1", IRanges( 50000000, 60000000 ) ), zlim=5 ) 
\dontrun{
print(pl)
}

}
\author{
Alejandro Reyes
}
